XRAYJENKINS-44 Minor refactoring
